Escape Keybinding (#282)
* Add standard menu keybinding for "Escape"

* fix pipeline
