+++ src/renderer/index.ts
@@ -5,7 +5,6 @@ import "@renderer/styles/styles.scss";
 
 import PrimeVue from "primevue/config";
 import Tooltip from "primevue/tooltip";
-import type { TransitionProps } from "vue";
 import { createApp } from "vue";
 import { createI18n } from "vue-i18n";
 import { localeFilePaths } from "@renderer/assets/assetFiles";
@@ -15,50 +14,27 @@ import { clickAwayDirective } from "@renderer/utils/click-away-directive";
 import { elementInViewDirective } from "@renderer/utils/element-in-view-directive";
 import { audioApi } from "@renderer/audio/audio";
 import { router } from "@renderer/router";
-import { settingsStore } from "@renderer/store/settings.store";
 import { initPreMountStores } from "@renderer/store/stores";
 
-declare module "vue-router" {
-    interface RouteMeta {
-        title?: string;
-        order?: number;
-        availableOffline?: boolean;
-        hide?: boolean;
-        empty?: boolean;
-        blurBg?: boolean;
-        transition?: TransitionProps;
-        overflowY?: "scroll" | "hidden";
-        devOnly?: boolean;
-        redirect?: string;
-    }
-}
-
-(async () => {
-    await setupVue();
-    window.addEventListener("keydown", (event) => {
-        if (event.code === "F11") {
-            event.preventDefault();
-            settingsStore.fullscreen = !settingsStore.fullscreen;
-        }
-    });
-})();
+setupVue();
 
 async function setupVue() {
     const app = createApp(App);
+
+    // Plugins
     app.use(router);
-    app.use(PrimeVue, {
-        ripple: true,
-    });
+    app.use(PrimeVue, { ripple: true });
     app.use(await setupI18n());
+
+    // Directives
     app.directive("click-away", clickAwayDirective);
     app.directive("in-view", elementInViewDirective);
     app.directive("tooltip", Tooltip);
-    if (process.env.NODE_ENV !== "production") {
-        app.config.globalProperties.window = window;
-    }
+
     // Init stores before mounting app
     await initPreMountStores();
     await audioApi.init();
+
     app.mount("#app");
 }
